[refactor] move on_reaction normalisation to the messenger-agnostic layer
Review feedback on #18: the emoji logic belongs in the framework proper, not
in the Telegram adaptor. Nothing in normalise_reaction_map is Telegram-specific
-- Emoji is already defined in messenger/context.py, and both Context and the
adaptor stub declare on_reaction in terms of it -- so it moves next to Emoji
and is imported by telegram/context.py.

Its tests move with it, from telegram/test_attach.py to messenger/test_context.py.

Pure move, no behaviour change. Full suite: 80 passed.
